backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
structlog==24.1.0
slowapi==0.1.9
groq==0.9.0
openai==1.30.1
faker==25.0.0
python-multipart==0.0.9
